move golden tests alongside relevant data
